Refs: update types using Fast DDS-Gen v2.3.0 and move to specific folder
