+++ types/KeylessShapeType.cxx
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 /*!
- * @file Shape.cpp
+ * @file KeylessShapeType.cpp
  * This source file contains the definition of the described types in the IDL file.
  *
  * This file was generated by the tool gen.
@@ -26,8 +26,8 @@ char dummy;
 }  // namespace
 #endif  // _WIN32
 
-#include "eprosimashapesdemo/shapesdemo/Shape.h"
-#include "eprosimashapesdemo/shapesdemo/ShapeTypeObject.h"
+#include "KeylessShapeType.h"
+#include "KeylessShapeTypeTypeObject.h"
 #include <fastcdr/Cdr.h>
 
 #include <fastcdr/exceptions/BadParamException.h>
@@ -35,22 +35,22 @@ using namespace eprosima::fastcdr::exception;
 
 #include <utility>
 
-ShapeType::ShapeType()
+shapes_demo_typesupport::idl::KeylessShapeType::KeylessShapeType()
 {
-    // m_color com.eprosima.idl.parser.typecode.StringTypeCode@5bfbf16f
+    // m_color com.eprosima.idl.parser.typecode.StringTypeCode@17d0685f
     m_color ="";
-    // m_x com.eprosima.idl.parser.typecode.PrimitiveTypeCode@25af5db5
+    // m_x com.eprosima.idl.parser.typecode.PrimitiveTypeCode@3891771e
     m_x = 0;
-    // m_y com.eprosima.idl.parser.typecode.PrimitiveTypeCode@12cdcf4
+    // m_y com.eprosima.idl.parser.typecode.PrimitiveTypeCode@78ac1102
     m_y = 0;
-    // m_shapesize com.eprosima.idl.parser.typecode.PrimitiveTypeCode@5bcea91b
+    // m_shapesize com.eprosima.idl.parser.typecode.PrimitiveTypeCode@2de8284b
     m_shapesize = 0;
 
     // Just to register all known types
-    // registerShapeTypes();
+    registerKeylessShapeTypeTypes();
 }
 
-ShapeType::~ShapeType()
+shapes_demo_typesupport::idl::KeylessShapeType::~KeylessShapeType()
 {
 
 
@@ -58,8 +58,8 @@ ShapeType::~ShapeType()
 
 }
 
-ShapeType::ShapeType(
-        const ShapeType& x)
+shapes_demo_typesupport::idl::KeylessShapeType::KeylessShapeType(
+        const KeylessShapeType& x)
 {
     m_color = x.m_color;
     m_x = x.m_x;
@@ -67,8 +67,8 @@ ShapeType::ShapeType(
     m_shapesize = x.m_shapesize;
 }
 
-ShapeType::ShapeType(
-        ShapeType&& x)
+shapes_demo_typesupport::idl::KeylessShapeType::KeylessShapeType(
+        KeylessShapeType&& x) noexcept 
 {
     m_color = std::move(x.m_color);
     m_x = x.m_x;
@@ -76,8 +76,8 @@ ShapeType::ShapeType(
     m_shapesize = x.m_shapesize;
 }
 
-ShapeType& ShapeType::operator =(
-        const ShapeType& x)
+shapes_demo_typesupport::idl::KeylessShapeType& shapes_demo_typesupport::idl::KeylessShapeType::operator =(
+        const KeylessShapeType& x)
 {
 
     m_color = x.m_color;
@@ -88,8 +88,8 @@ ShapeType& ShapeType::operator =(
     return *this;
 }
 
-ShapeType& ShapeType::operator =(
-        ShapeType&& x)
+shapes_demo_typesupport::idl::KeylessShapeType& shapes_demo_typesupport::idl::KeylessShapeType::operator =(
+        KeylessShapeType&& x) noexcept
 {
 
     m_color = std::move(x.m_color);
@@ -100,20 +100,20 @@ ShapeType& ShapeType::operator =(
     return *this;
 }
 
-bool ShapeType::operator ==(
-        const ShapeType& x) const
+bool shapes_demo_typesupport::idl::KeylessShapeType::operator ==(
+        const KeylessShapeType& x) const
 {
 
     return (m_color == x.m_color && m_x == x.m_x && m_y == x.m_y && m_shapesize == x.m_shapesize);
 }
 
-bool ShapeType::operator !=(
-        const ShapeType& x) const
+bool shapes_demo_typesupport::idl::KeylessShapeType::operator !=(
+        const KeylessShapeType& x) const
 {
     return !(*this == x);
 }
 
-size_t ShapeType::getMaxCdrSerializedSize(
+size_t shapes_demo_typesupport::idl::KeylessShapeType::getMaxCdrSerializedSize(
         size_t current_alignment)
 {
     size_t initial_alignment = current_alignment;
@@ -134,8 +134,8 @@ size_t ShapeType::getMaxCdrSerializedSize(
     return current_alignment - initial_alignment;
 }
 
-size_t ShapeType::getCdrSerializedSize(
-        const ShapeType& data,
+size_t shapes_demo_typesupport::idl::KeylessShapeType::getCdrSerializedSize(
+        const shapes_demo_typesupport::idl::KeylessShapeType& data,
         size_t current_alignment)
 {
     (void)data;
@@ -157,18 +157,18 @@ size_t ShapeType::getCdrSerializedSize(
     return current_alignment - initial_alignment;
 }
 
-void ShapeType::serialize(
+void shapes_demo_typesupport::idl::KeylessShapeType::serialize(
         eprosima::fastcdr::Cdr& scdr) const
 {
 
-    scdr << m_color;
+    scdr << m_color.c_str();
     scdr << m_x;
     scdr << m_y;
     scdr << m_shapesize;
 
 }
 
-void ShapeType::deserialize(
+void shapes_demo_typesupport::idl::KeylessShapeType::deserialize(
         eprosima::fastcdr::Cdr& dcdr)
 {
 
@@ -182,7 +182,7 @@ void ShapeType::deserialize(
  * @brief This function copies the value in member color
  * @param _color New value to be copied in member color
  */
-void ShapeType::color(
+void shapes_demo_typesupport::idl::KeylessShapeType::color(
         const std::string& _color)
 {
     m_color = _color;
@@ -192,7 +192,7 @@ void ShapeType::color(
  * @brief This function moves the value in member color
  * @param _color New value to be moved in member color
  */
-void ShapeType::color(
+void shapes_demo_typesupport::idl::KeylessShapeType::color(
         std::string&& _color)
 {
     m_color = std::move(_color);
@@ -202,7 +202,7 @@ void ShapeType::color(
  * @brief This function returns a constant reference to member color
  * @return Constant reference to member color
  */
-const std::string& ShapeType::color() const
+const std::string& shapes_demo_typesupport::idl::KeylessShapeType::color() const
 {
     return m_color;
 }
@@ -211,7 +211,7 @@ const std::string& ShapeType::color() const
  * @brief This function returns a reference to member color
  * @return Reference to member color
  */
-std::string& ShapeType::color()
+std::string& shapes_demo_typesupport::idl::KeylessShapeType::color()
 {
     return m_color;
 }
@@ -219,7 +219,7 @@ std::string& ShapeType::color()
  * @brief This function sets a value in member x
  * @param _x New value for member x
  */
-void ShapeType::x(
+void shapes_demo_typesupport::idl::KeylessShapeType::x(
         int32_t _x)
 {
     m_x = _x;
@@ -229,7 +229,7 @@ void ShapeType::x(
  * @brief This function returns the value of member x
  * @return Value of member x
  */
-int32_t ShapeType::x() const
+int32_t shapes_demo_typesupport::idl::KeylessShapeType::x() const
 {
     return m_x;
 }
@@ -238,7 +238,7 @@ int32_t ShapeType::x() const
  * @brief This function returns a reference to member x
  * @return Reference to member x
  */
-int32_t& ShapeType::x()
+int32_t& shapes_demo_typesupport::idl::KeylessShapeType::x()
 {
     return m_x;
 }
@@ -247,7 +247,7 @@ int32_t& ShapeType::x()
  * @brief This function sets a value in member y
  * @param _y New value for member y
  */
-void ShapeType::y(
+void shapes_demo_typesupport::idl::KeylessShapeType::y(
         int32_t _y)
 {
     m_y = _y;
@@ -257,7 +257,7 @@ void ShapeType::y(
  * @brief This function returns the value of member y
  * @return Value of member y
  */
-int32_t ShapeType::y() const
+int32_t shapes_demo_typesupport::idl::KeylessShapeType::y() const
 {
     return m_y;
 }
@@ -266,7 +266,7 @@ int32_t ShapeType::y() const
  * @brief This function returns a reference to member y
  * @return Reference to member y
  */
-int32_t& ShapeType::y()
+int32_t& shapes_demo_typesupport::idl::KeylessShapeType::y()
 {
     return m_y;
 }
@@ -275,7 +275,7 @@ int32_t& ShapeType::y()
  * @brief This function sets a value in member shapesize
  * @param _shapesize New value for member shapesize
  */
-void ShapeType::shapesize(
+void shapes_demo_typesupport::idl::KeylessShapeType::shapesize(
         int32_t _shapesize)
 {
     m_shapesize = _shapesize;
@@ -285,7 +285,7 @@ void ShapeType::shapesize(
  * @brief This function returns the value of member shapesize
  * @return Value of member shapesize
  */
-int32_t ShapeType::shapesize() const
+int32_t shapes_demo_typesupport::idl::KeylessShapeType::shapesize() const
 {
     return m_shapesize;
 }
@@ -294,20 +294,19 @@ int32_t ShapeType::shapesize() const
  * @brief This function returns a reference to member shapesize
  * @return Reference to member shapesize
  */
-int32_t& ShapeType::shapesize()
+int32_t& shapes_demo_typesupport::idl::KeylessShapeType::shapesize()
 {
     return m_shapesize;
 }
 
 
-size_t ShapeType::getKeyMaxCdrSerializedSize(
+size_t shapes_demo_typesupport::idl::KeylessShapeType::getKeyMaxCdrSerializedSize(
         size_t current_alignment)
 {
     size_t current_align = current_alignment;
 
 
-     current_align += 4 + eprosima::fastcdr::Cdr::alignment(current_align, 4) + 255 + 1;
-     
+
 
 
 
@@ -315,15 +314,16 @@ size_t ShapeType::getKeyMaxCdrSerializedSize(
     return current_align;
 }
 
-bool ShapeType::isKeyDefined()
+bool shapes_demo_typesupport::idl::KeylessShapeType::isKeyDefined()
 {
-    return true;
+    return false;
 }
 
-void ShapeType::serializeKey(
+void shapes_demo_typesupport::idl::KeylessShapeType::serializeKey(
         eprosima::fastcdr::Cdr& scdr) const
 {
     (void) scdr;
-     scdr << m_color;
-         
+        
 }
+
+
