+++ types/KeylessShapeType.h
@@ -13,15 +13,17 @@
 // limitations under the License.
 
 /*!
- * @file KeylessShape.h
+ * @file KeylessShapeType.h
  * This header file contains the declaration of the described types in the IDL file.
  *
  * This file was generated by the tool gen.
  */
 
-#ifndef _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_H_
-#define _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_H_
+#ifndef _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_H_
+#define _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_H_
 
+
+#include <fastrtps/utils/fixed_size_string.hpp>
 
 #include <stdint.h>
 #include <array>
@@ -42,16 +44,16 @@
 
 #if defined(_WIN32)
 #if defined(EPROSIMA_USER_DLL_EXPORT)
-#if defined(KeylessShape_SOURCE)
-#define KeylessShape_DllAPI __declspec( dllexport )
+#if defined(KeylessShapeType_SOURCE)
+#define KeylessShapeType_DllAPI __declspec( dllexport )
 #else
-#define KeylessShape_DllAPI __declspec( dllimport )
-#endif // KeylessShape_SOURCE
+#define KeylessShapeType_DllAPI __declspec( dllimport )
+#endif // KeylessShapeType_SOURCE
 #else
-#define KeylessShape_DllAPI
+#define KeylessShapeType_DllAPI
 #endif  // EPROSIMA_USER_DLL_EXPORT
 #else
-#define KeylessShape_DllAPI
+#define KeylessShapeType_DllAPI
 #endif // _WIN32
 
 namespace eprosima {
@@ -65,7 +67,7 @@ namespace shapes_demo_typesupport {
     namespace idl {
         /*!
          * @brief This class represents the structure KeylessShapeType defined by the user in the IDL file.
-         * @ingroup KEYLESSSHAPE
+         * @ingroup KEYLESSSHAPETYPE
          */
         class KeylessShapeType
         {
@@ -93,7 +95,7 @@ namespace shapes_demo_typesupport {
              * @param x Reference to the object shapes_demo_typesupport::idl::KeylessShapeType that will be copied.
              */
             eProsima_user_DllExport KeylessShapeType(
-                    KeylessShapeType&& x);
+                    KeylessShapeType&& x) noexcept;
 
             /*!
              * @brief Copy assignment.
@@ -107,7 +109,7 @@ namespace shapes_demo_typesupport {
              * @param x Reference to the object shapes_demo_typesupport::idl::KeylessShapeType that will be copied.
              */
             eProsima_user_DllExport KeylessShapeType& operator =(
-                    KeylessShapeType&& x);
+                    KeylessShapeType&& x) noexcept;
 
             /*!
              * @brief Comparison operator.
@@ -273,4 +275,4 @@ namespace shapes_demo_typesupport {
     } // namespace idl
 } // namespace shapes_demo_typesupport
 
-#endif // _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_H_
+#endif // _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_H_
