+++ types/KeylessShapeTypePubSubTypes.cxx
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 /*!
- * @file KeylessShapePubSubTypes.cpp
+ * @file KeylessShapeTypePubSubTypes.cpp
  * This header file contains the implementation of the serialization functions.
  *
  * This file was generated by the tool fastcdrgen.
@@ -23,7 +23,7 @@
 #include <fastcdr/FastBuffer.h>
 #include <fastcdr/Cdr.h>
 
-#include "eprosimashapesdemo/shapesdemo/KeylessShapePubSubTypes.h"
+#include "KeylessShapeTypePubSubTypes.h"
 
 using SerializedPayload_t = eprosima::fastrtps::rtps::SerializedPayload_t;
 using InstanceHandle_t = eprosima::fastrtps::rtps::InstanceHandle_t;
@@ -32,7 +32,7 @@ namespace shapes_demo_typesupport {
     namespace idl {
         KeylessShapeTypePubSubType::KeylessShapeTypePubSubType()
         {
-            setName("shapes_demo_typesupport::idl::dds_::KeylessShapeType_");
+            setName("shapes_demo_typesupport::idl::KeylessShapeType");
             auto type_size = KeylessShapeType::getMaxCdrSerializedSize();
             type_size += eprosima::fastcdr::Cdr::alignment(type_size, 4); /* possible submessage alignment */
             m_typeSize = static_cast<uint32_t>(type_size) + 4; /*encapsulation*/
