+++ types/KeylessShapeTypePubSubTypes.h
@@ -13,24 +13,24 @@
 // limitations under the License.
 
 /*!
- * @file KeylessShapePubSubTypes.h
+ * @file KeylessShapeTypePubSubTypes.h
  * This header file contains the declaration of the serialization functions.
  *
  * This file was generated by the tool fastcdrgen.
  */
 
 
-#ifndef _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_PUBSUBTYPES_H_
-#define _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_PUBSUBTYPES_H_
+#ifndef _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_PUBSUBTYPES_H_
+#define _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_PUBSUBTYPES_H_
 
 #include <fastdds/dds/topic/TopicDataType.hpp>
 #include <fastrtps/utils/md5.h>
 
-#include "eprosimashapesdemo/shapesdemo/KeylessShape.h"
+#include "KeylessShapeType.h"
 
 #if !defined(GEN_API_VER) || (GEN_API_VER != 1)
 #error \
-    Generated KeylessShape is not compatible with current installed Fast DDS. Please, regenerate it with fastddsgen.
+    Generated KeylessShapeType is not compatible with current installed Fast DDS. Please, regenerate it with fastddsgen.
 #endif  // GEN_API_VER
 
 namespace shapes_demo_typesupport
@@ -39,7 +39,7 @@ namespace shapes_demo_typesupport
     {
         /*!
          * @brief This class represents the TopicDataType of the type KeylessShapeType defined by the user in the IDL file.
-         * @ingroup KEYLESSSHAPE
+         * @ingroup KEYLESSSHAPETYPE
          */
         class KeylessShapeTypePubSubType : public eprosima::fastdds::dds::TopicDataType
         {
@@ -49,7 +49,7 @@ namespace shapes_demo_typesupport
 
             eProsima_user_DllExport KeylessShapeTypePubSubType();
 
-            eProsima_user_DllExport virtual ~KeylessShapeTypePubSubType();
+            eProsima_user_DllExport virtual ~KeylessShapeTypePubSubType() override;
 
             eProsima_user_DllExport virtual bool serialize(
                     void* data,
@@ -104,4 +104,4 @@ namespace shapes_demo_typesupport
     }
 }
 
-#endif // _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_PUBSUBTYPES_H_
+#endif // _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_PUBSUBTYPES_H_
