+++ types/KeylessShapeTypeTypeObject.cxx
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 /*!
- * @file KeylessShapeTypeObject.cpp
+ * @file KeylessShapeTypeTypeObject.cpp
  * This source file contains the definition of the described types in the IDL file.
  *
  * This file was generated by the tool gen.
@@ -24,8 +24,8 @@
 namespace { char dummy; }
 #endif
 
-#include "eprosimashapesdemo/shapesdemo/KeylessShape.h"
-#include "eprosimashapesdemo/shapesdemo/KeylessShapeTypeObject.h"
+#include "KeylessShapeType.h"
+#include "KeylessShapeTypeTypeObject.h"
 #include <utility>
 #include <sstream>
 #include <fastrtps/rtps/common/SerializedPayload.h>
@@ -38,12 +38,12 @@ namespace { char dummy; }
 
 using namespace eprosima::fastrtps::rtps;
 
-void registerKeylessShapeTypes()
+void registerKeylessShapeTypeTypes()
 {
     TypeObjectFactory *factory = TypeObjectFactory::get_instance();
-    factory->add_type_object("shapes_demo_typesupport::idl::dds_::KeylessShapeType_", shapes_demo_typesupport::idl::GetKeylessShapeTypeIdentifier(true),
+    factory->add_type_object("shapes_demo_typesupport::idl::KeylessShapeType", shapes_demo_typesupport::idl::GetKeylessShapeTypeIdentifier(true),
             shapes_demo_typesupport::idl::GetKeylessShapeTypeObject(true));
-    factory->add_type_object("shapes_demo_typesupport::idl::dds_::KeylessShapeType_", shapes_demo_typesupport::idl::GetKeylessShapeTypeIdentifier(false),
+    factory->add_type_object("shapes_demo_typesupport::idl::KeylessShapeType", shapes_demo_typesupport::idl::GetKeylessShapeTypeIdentifier(false),
             shapes_demo_typesupport::idl::GetKeylessShapeTypeObject(false));
