+++ types/KeylessShapeTypeTypeObject.h
@@ -13,14 +13,14 @@
 // limitations under the License.
 
 /*!
- * @file KeylessShapeTypeObject.h
+ * @file KeylessShapeTypeTypeObject.h
  * This header file contains the declaration of the described types in the IDL file.
  *
  * This file was generated by the tool gen.
  */
 
-#ifndef _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_TYPE_OBJECT_H_
-#define _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_TYPE_OBJECT_H_
+#ifndef _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_TYPE_OBJECT_H_
+#define _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_TYPE_OBJECT_H_
 
 
 #include <fastrtps/types/TypeObject.h>
@@ -38,21 +38,21 @@
 
 #if defined(_WIN32)
 #if defined(EPROSIMA_USER_DLL_EXPORT)
-#if defined(KeylessShape_SOURCE)
-#define KeylessShape_DllAPI __declspec( dllexport )
+#if defined(KeylessShapeType_SOURCE)
+#define KeylessShapeType_DllAPI __declspec( dllexport )
 #else
-#define KeylessShape_DllAPI __declspec( dllimport )
-#endif // KeylessShape_SOURCE
+#define KeylessShapeType_DllAPI __declspec( dllimport )
+#endif // KeylessShapeType_SOURCE
 #else
-#define KeylessShape_DllAPI
+#define KeylessShapeType_DllAPI
 #endif
 #else
-#define KeylessShape_DllAPI
+#define KeylessShapeType_DllAPI
 #endif // _WIN32
 
 using namespace eprosima::fastrtps::types;
 
-eProsima_user_DllExport void registerKeylessShapeTypes();
+eProsima_user_DllExport void registerKeylessShapeTypeTypes();
 
 namespace shapes_demo_typesupport {
     namespace idl {
@@ -64,4 +64,4 @@ namespace shapes_demo_typesupport {
     } // namespace idl
 } // namespace shapes_demo_typesupport
 
-#endif // _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPE_TYPE_OBJECT_H_
+#endif // _FAST_DDS_GENERATED_SHAPES_DEMO_TYPESUPPORT_IDL_KEYLESSSHAPETYPE_TYPE_OBJECT_H_
