+++ types/ShapeTypeObject.h
